+++ lib/rspec/core/hooks.rb
@@ -448,6 +448,11 @@ module RSpec
                             "`#{position}(:suite)` hook, registered on an example " \
                             "group, will be ignored."
             return
+          elsif scope == :context && position == :around
+            # TODO: consider making this an error in RSpec 4. For SemVer reasons,
+            # we are only warning in RSpec 3.
+            RSpec.warn_with "WARNING: `around(:context)` hooks are not supported and " \
+                            "behave like `around(:example)."
           end
 
           hook = HOOK_TYPES[position][scope].new(block, options)